@@ -8,18 +8,18 @@
 import Foundation
 
 final class Context {
-	
-	let environment: Environment
-	let keychain: KeychainStorage
-	let network: Network
-	let cache: Cache
-	
-	init(environment: Environment) {
-		self.environment = environment
-		self.keychain = KeychainService()
-		self.network = Network(environment: environment, keychain: keychain)
-		self.cache = Cache(environment: environment)
-	}
-	
+    
+    let environment: Environment
+    let keychain: KeychainStorage
+    let network: Network
+    let cache: Cache
+    
+    init(environment: Environment) {
+        self.environment = environment
+        self.keychain = KeychainService()
+        self.network = Network(environment: environment, keychain: keychain)
+        self.cache = Cache(environment: environment)
+    }
+    
 }
 
