@@ -8,16 +8,12 @@
 import Foundation
 
 final class Environment {
-	var baseURL: URL {
-		guard let url = URL(string: "https://api.monobank.ua/") else {
-			fatalError("Failed to setup invalid URL")
-		}
-		return url
-	}
-	
-	var authorizedCacheDirectoryName: String { "MonobankAuhtorized" }
-	var unauthorizedCacheDirectoryName: String { "MonobankUnauhtorized" }
-	
-	var authorizedCacheExpiry: TimeInterval { 60 }
-	var unauthorizedCacheExpiry: TimeInterval { 60*5 }
+    var baseURL: URL {
+        guard let url = URL(string: "https://api.monobank.ua/") else {
+            fatalError("Failed to setup invalid URL")
+        }
+        return url
+    }
+    var authorizedCacheExpiry: TimeInterval { 60 }
+    var unauthorizedCacheExpiry: TimeInterval { 60*5 }
 }
