@@ -9,39 +9,39 @@ import Foundation
 import KeychainAccess
 
 protocol KeychainStorage {
-	func value(for key: String) -> String?
-	func set(value: String, for key: String)
-	func removeValue(for key: String)
-	func clear()
+    func value(for key: String) -> String?
+    func set(value: String, for key: String)
+    func removeValue(for key: String)
+    func clear()
 }
 
 final class KeychainService: KeychainStorage {
-	
-	private let keychain: Keychain
-	
-	init() {
-		self.keychain = Keychain()
-	}
-	
-	func value(for key: String) -> String? {
-		keychain[key]
-	}
-	
-	func set(value: String, for key: String) {
-		keychain[key] = value
-	}
-	
-	func removeValue(for key: String) {
-		keychain[key] = nil
-	}
-	
-	func clear() {
-		try? keychain.removeAll()
-	}
+    
+    private let keychain: Keychain
+    
+    init() {
+        self.keychain = Keychain()
+    }
+    
+    func value(for key: String) -> String? {
+        keychain[key]
+    }
+    
+    func set(value: String, for key: String) {
+        keychain[key] = value
+    }
+    
+    func removeValue(for key: String) {
+        keychain[key] = nil
+    }
+    
+    func clear() {
+        try? keychain.removeAll()
+    }
 }
 
 extension KeychainService {
-	enum Key: String {
-		case token
-	}
+    enum Key: String {
+        case token
+    }
 }
