@@ -11,8 +11,7 @@ import Foundation
 protocol API: TargetType { }
 
 extension API {
-	var baseURL: URL { fatalError("BaseURL configured in Network core class") }
-	
-	var method: Moya.Method { .get }
+    var baseURL: URL { fatalError("BaseURL configured in Network core class") }
+    var method: Moya.Method { .get }
 }
 
