@@ -9,17 +9,17 @@ import Foundation
 import Moya
 
 enum ClientInfoAPI: API {
-	
-	case personalInfo
-	
-	var path: String { "/personal/client-info" }
-	
-	var sampleData: Data { path.data(using: .utf8) ?? Data() }
-	
-	var task: Task { .requestPlain }
-	
-	var headers: [String : String]? { nil }
-	
-	var needAuth: Bool { true }
+    
+    case personalInfo
+    
+    var path: String { "/personal/client-info" }
+    
+    var sampleData: Data { path.data(using: .utf8) ?? Data() }
+    
+    var task: Task { .requestPlain }
+    
+    var headers: [String : String]? { nil }
+    
+    var needAuth: Bool { true }
 }
 
